requirements.txt
flask==3.0.0
gunicorn==21.2.0
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
PyYAML==6.0.2
tenacity==8.5.0
